// src/app/lib/types/shared/alert.types.ts
export type GnxAlertAppearance =
  | 'border'
  | 'fill'
  | 'outline'
  | 'soft';

export type GnxAlertType =
  | 'primary'
  | 'accent'
  | 'warn'
  | 'basic'
  | 'info'
  | 'success'
  | 'warning'
  | 'error';

export type GnxAlertAutoCloseDuration =
  | 'short'
  | 'medium'
  | 'long'

export type GnxAlertLevel =
| 'header'
| 'page'
| 'footer'
| 'form'